api/requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy-libsql
aiosqlite
python-dotenv
sqlite-utils
# ...existing code...